refactor(cli): extract driver/core inline test modules to stay under size ratchet
core.rs exceeded the 3195-line monolith ratchet (was 3404 after the
merge-cache addition). Extract two inline test modules to separate files
following the existing pattern of #[path = "..."] mod declarations:

- explain_files_reason_tests → explain_files_reason_tests.rs (59 lines)
- merge_cache_tests         → core_merge_cache_tests.rs     (171 lines)

core.rs is now 3179 lines, 16 under the ratchet. All 8 tests in the
extracted modules continue to pass.
