+++ crates/tsz-cli/src/driver/core.rs
@@ -3171,234 +3171,9 @@ mod tests;
 mod cross_file_circular_alias_tests;
 
 #[cfg(test)]
-mod explain_files_reason_tests {
-    use super::*;
-
-    /// Issue #3901: tsc surfaces tsconfig `files` entries with a distinct
-    /// reason from `include` matches.
-    #[test]
-    fn files_list_entry_renders_tsc_phrasing() {
-        assert_eq!(
-            FileInclusionReason::FilesListEntry.to_string(),
-            "Part of 'files' list in tsconfig.json"
-        );
-    }
-
-    /// Default-lib reasons must mention the configured target so users
-    /// can attribute the lib pull. tsc renders the lowercase ECMAScript
-    /// revision name.
-    #[test]
-    fn default_library_reason_includes_target() {
-        assert_eq!(
-            FileInclusionReason::DefaultLibrary("es2018".to_string()).to_string(),
-            "Default library for target 'es2018'"
-        );
-    }
-
-    /// `is_default_lib_for_target` matches both the `lib.<target>.full.d.ts`
-    /// and `lib.<target>.d.ts` shapes that the lib resolver produces.
-    #[test]
-    fn default_lib_matches_full_and_bare_for_target() {
-        let full = PathBuf::from("/usr/typescript/lib.es2018.full.d.ts");
-        let bare = PathBuf::from("/usr/typescript/lib.es2018.d.ts");
-        let other_target = PathBuf::from("/usr/typescript/lib.es2020.full.d.ts");
-        let unrelated = PathBuf::from("/usr/typescript/lib.dom.d.ts");
-
-        assert!(is_default_lib_for_target(&full, ScriptTarget::ES2018));
-        assert!(is_default_lib_for_target(&bare, ScriptTarget::ES2018));
-        assert!(!is_default_lib_for_target(
-            &other_target,
-            ScriptTarget::ES2018
-        ));
-        assert!(!is_default_lib_for_target(&unrelated, ScriptTarget::ES2018));
-    }
-
-    /// Locks the lowercase target spelling for the explainFiles surface.
-    #[test]
-    fn target_display_for_explain_files_lowercase() {
-        assert_eq!(
-            script_target_display_for_explain_files(ScriptTarget::ES5),
-            "es5"
-        );
-        assert_eq!(
-            script_target_display_for_explain_files(ScriptTarget::ES2018),
-            "es2018"
-        );
-        assert_eq!(
-            script_target_display_for_explain_files(ScriptTarget::ESNext),
-            "esnext"
-        );
-    }
-}
+#[path = "explain_files_reason_tests.rs"]
+mod explain_files_reason_tests;
 
 #[cfg(test)]
-mod merge_cache_tests {
-    use super::sources::SourceEntry;
-    use super::*;
-
-    fn make_source(path: &str, text: &str) -> SourceEntry {
-        SourceEntry {
-            path: PathBuf::from(path),
-            text: Some(text.to_string()),
-            is_binary: false,
-            suppress_parser_diagnostics: false,
-        }
-    }
-
-    fn fresh_cache() -> (CompilationCache, Vec<Arc<LibFile>>) {
-        (CompilationCache::default(), vec![])
-    }
-
-    /// First call with an empty cache always runs the merge.
-    /// Second call with the same unchanged inputs returns the same `Arc`
-    /// pointer (no re-merge). Third call after a file change re-merges and
-    /// produces a new pointer.
-    #[test]
-    fn merge_skipped_on_unchanged_inputs_and_reruns_on_change() {
-        let (mut cache, libs) = fresh_cache();
-
-        // --- first call: cold cache, must merge ---
-        let sources = vec![
-            make_source("/a.ts", "export const a = 1;"),
-            make_source("/b.ts", "export const b = 2;"),
-        ];
-        let r1 = build_program_with_cache(sources, &mut cache, &libs, ScriptTarget::ES2020);
-        let ptr1 = Arc::as_ptr(&r1.program);
-        assert!(
-            !r1.dirty_paths.is_empty(),
-            "first build should have dirty paths"
-        );
-
-        // --- second call: identical inputs, merge must be skipped ---
-        let sources = vec![
-            make_source("/a.ts", "export const a = 1;"),
-            make_source("/b.ts", "export const b = 2;"),
-        ];
-        let r2 = build_program_with_cache(sources, &mut cache, &libs, ScriptTarget::ES2020);
-        let ptr2 = Arc::as_ptr(&r2.program);
-        assert!(
-            r2.dirty_paths.is_empty(),
-            "second build with unchanged inputs should have no dirty paths"
-        );
-        assert_eq!(
-            ptr1, ptr2,
-            "unchanged inputs must return the same Arc<MergedProgram> pointer"
-        );
-
-        // --- third call: one file changed, must re-merge ---
-        let sources = vec![
-            make_source("/a.ts", "export const a = 99;"), // changed
-            make_source("/b.ts", "export const b = 2;"),
-        ];
-        let r3 = build_program_with_cache(sources, &mut cache, &libs, ScriptTarget::ES2020);
-        let ptr3 = Arc::as_ptr(&r3.program);
-        assert!(
-            !r3.dirty_paths.is_empty(),
-            "modified-file build should have dirty paths"
-        );
-        assert_ne!(
-            ptr2, ptr3,
-            "changed inputs must produce a new Arc<MergedProgram>"
-        );
-
-        // --- fourth call: back to unchanged after change, cache is warm again ---
-        let sources = vec![
-            make_source("/a.ts", "export const a = 99;"),
-            make_source("/b.ts", "export const b = 2;"),
-        ];
-        let r4 = build_program_with_cache(sources, &mut cache, &libs, ScriptTarget::ES2020);
-        assert!(
-            r4.dirty_paths.is_empty(),
-            "fourth build (re-stable) should have no dirty paths"
-        );
-        assert_eq!(
-            Arc::as_ptr(&r3.program),
-            Arc::as_ptr(&r4.program),
-            "re-stable inputs must return the same Arc as the last merge"
-        );
-    }
-
-    /// Removing a file invalidates the cached merge (file-count guard).
-    #[test]
-    fn merge_invalidated_on_file_removal() {
-        let (mut cache, libs) = fresh_cache();
-
-        let sources = vec![
-            make_source("/a.ts", "export const a = 1;"),
-            make_source("/b.ts", "export const b = 2;"),
-        ];
-        let r1 = build_program_with_cache(sources, &mut cache, &libs, ScriptTarget::ES2020);
-
-        // Warm the cache with a no-op second pass.
-        let sources = vec![
-            make_source("/a.ts", "export const a = 1;"),
-            make_source("/b.ts", "export const b = 2;"),
-        ];
-        let r2 = build_program_with_cache(sources, &mut cache, &libs, ScriptTarget::ES2020);
-        assert_eq!(Arc::as_ptr(&r1.program), Arc::as_ptr(&r2.program));
-
-        // Now remove one file.
-        let sources = vec![make_source("/a.ts", "export const a = 1;")];
-        let r3 = build_program_with_cache(sources, &mut cache, &libs, ScriptTarget::ES2020);
-        assert_ne!(
-            Arc::as_ptr(&r2.program),
-            Arc::as_ptr(&r3.program),
-            "file removal must trigger a fresh merge"
-        );
-    }
-
-    /// Adding a file invalidates the cached merge.
-    #[test]
-    fn merge_invalidated_on_file_addition() {
-        let (mut cache, libs) = fresh_cache();
-
-        let sources = vec![make_source("/a.ts", "export const a = 1;")];
-        let r1 = build_program_with_cache(sources, &mut cache, &libs, ScriptTarget::ES2020);
-
-        // Warm the cache.
-        let sources = vec![make_source("/a.ts", "export const a = 1;")];
-        let r2 = build_program_with_cache(sources, &mut cache, &libs, ScriptTarget::ES2020);
-        assert_eq!(Arc::as_ptr(&r1.program), Arc::as_ptr(&r2.program));
-
-        // Add a new file.
-        let sources = vec![
-            make_source("/a.ts", "export const a = 1;"),
-            make_source("/c.ts", "export const c = 3;"),
-        ];
-        let r3 = build_program_with_cache(sources, &mut cache, &libs, ScriptTarget::ES2020);
-        assert_ne!(
-            Arc::as_ptr(&r2.program),
-            Arc::as_ptr(&r3.program),
-            "file addition must trigger a fresh merge"
-        );
-    }
-
-    /// `CompilationCache::clear()` removes the cached merged program.
-    #[test]
-    fn clear_invalidates_merge_cache() {
-        let (mut cache, libs) = fresh_cache();
-
-        let sources = vec![make_source("/a.ts", "export const a = 1;")];
-        build_program_with_cache(sources, &mut cache, &libs, ScriptTarget::ES2020);
-
-        // Warm the merge cache.
-        let sources = vec![make_source("/a.ts", "export const a = 1;")];
-        let r2 = build_program_with_cache(sources, &mut cache, &libs, ScriptTarget::ES2020);
-        assert!(r2.dirty_paths.is_empty());
-
-        cache.clear();
-
-        // After clear, next build must re-merge.
-        let sources = vec![make_source("/a.ts", "export const a = 1;")];
-        let r3 = build_program_with_cache(sources, &mut cache, &libs, ScriptTarget::ES2020);
-        assert!(
-            !r3.dirty_paths.is_empty(),
-            "first build after clear must re-parse and re-merge"
-        );
-        assert_ne!(
-            Arc::as_ptr(&r2.program),
-            Arc::as_ptr(&r3.program),
-            "clear must produce a fresh Arc<MergedProgram>"
-        );
-    }
-}
+#[path = "core_merge_cache_tests.rs"]
+mod merge_cache_tests;
